+++ Normalization.h
@@ -16,6 +16,6 @@ using namespace std;
 /******************************************************************************
 								Function Prototypes
 ******************************************************************************/
-void normalize_pdsi( vector< vector<float> >& data );
-void normalize_burned_acres( vector< vector<float> >& data );
+void normalize_pdsi( vector< vector<double> >& data );
+void normalize_burned_acres( vector< vector<double> >& data );
